# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Personal household budget API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /settings:
    get:
      operationId: getSettings
      tags: [settings]
      summary: Get app settings
      responses:
        "200":
          description: Settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Settings"
    put:
      operationId: updateSettings
      tags: [settings]
      summary: Update app settings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SettingsUpdate"
      responses:
        "200":
          description: Updated settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Settings"
...
  /incomes:
    get:
      operationId: listIncomes
      tags: [incomes]
      summary: List income sources
      responses:
        "200":
          description: Income sources
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/IncomeSource"
    post:
      operationId: createIncome
      tags: [incomes]
      summary: Create income source
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IncomeSourceInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IncomeSource"
  /incomes/{id}:
    patch:
      operationId: updateIncome
      tags: [incomes]
      summary: Update income source
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IncomeSourceUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IncomeSource"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiMessage"
    delete:
      operationId: deleteIncome
      tags: [incomes]
      summary: Delete income source
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiMessage"
  /categories:
    get:
      operationId: listCategories
      tags: [categories]
      summary: List budget categories with their subcategories
      responses:
        "200":
          description: Categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"
    post:
      operationId: createCategory
      tags: [categories]
      summary: Create category
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
  /categories/{id}:
    patch:
      operationId: updateCategory
      tags: [categories]
      summary: Update category (rename or edit subcategory list)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiMessage"
    delete:
      operationId: deleteCategory
      tags: [categories]
      summary: Delete category
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiMessage"
  /plan:
    get:
      operationId: listPlanLines
      tags: [plan]
      summary: List budget plan lines
      responses:
        "200":
          description: Plan lines
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PlanLine"
    post:
      operationId: createPlanLine
      tags: [plan]
      summary: Create budget plan line
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlanLineInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlanLine"
  /plan/{id}:
    patch:
      operationId: updatePlanLine
      tags: [plan]
      summary: Update budget plan line
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlanLineUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlanLine"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiMessage"
    delete:
      operationId: deletePlanLine
      tags: [plan]
      summary: Delete budget plan line
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiMessage"
...
  /rules:
    get:
      operationId: listRules
      tags: [rules]
      summary: List merchant categorization rules
      responses:
        "200":
          description: Rules
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Rule"
    post:
      operationId: createRule
      tags: [rules]
      summary: Create a merchant rule (retroactively applied to unreviewed transactions)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RuleInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Rule"
  /rules/{id}:
    delete:
      operationId: deleteRule
      tags: [rules]
      summary: Delete a merchant rule
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiMessage"
  /dashboard:
    get:
      operationId: getDashboard
      tags: [dashboard]
      summary: Dashboard summary for a budget month
      parameters:
        - name: month
          in: query
          required: false
          schema: { type: string }
      responses:
        "200":
          description: Dashboard summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardSummary"
  /reconciliation:
    get:
      operationId: getReconciliation
      tags: [reconciliation]
      summary: Manual vs bank actuals comparison per category for a budget month
      parameters:
        - name: month
          in: query
          required: false
          schema: { type: string }
      responses:
        "200":
          description: Reconciliation rows
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ReconciliationRow"
...
components:
  schemas:
...
    ApiMessage:
      type: object
      required: [error]
      properties:
        error: { type: string }
    Settings:
      type: object
      required: [selectedMonth, monthStartDay, checkingBuffer, debtStrategy]
      properties:
        selectedMonth: { type: string }
        monthStartDay: { type: number }
        checkingBuffer: { type: number }
        debtStrategy: { type: string }
    SettingsUpdate:
      type: object
      properties:
        selectedMonth: { type: string }
        monthStartDay: { type: number, minimum: 1, maximum: 28 }
        checkingBuffer: { type: number }
        debtStrategy: { type: string }
    IncomeSource:
      type: object
      required: [id, name, frequency, netAmount, monthlyEquivalent]
      properties:
        id: { type: number }
        name: { type: string }
        owner: { type: ["string", "null"] }
        frequency: { type: string }
        netAmount: { type: number }
        monthlyEquivalent: { type: number }
        notes: { type: ["string", "null"] }
    IncomeSourceInput:
      type: object
      required: [name, frequency, netAmount, monthlyEquivalent]
      properties:
        name: { type: string, minLength: 1 }
        owner: { type: ["string", "null"] }
        frequency: { type: string }
        netAmount: { type: number }
        monthlyEquivalent: { type: number }
        notes: { type: ["string", "null"] }
    IncomeSourceUpdate:
      type: object
      properties:
        name: { type: string, minLength: 1 }
        owner: { type: ["string", "null"] }
        frequency: { type: string }
        netAmount: { type: number }
        monthlyEquivalent: { type: number }
        notes: { type: ["string", "null"] }
    Category:
      type: object
      required: [id, name, subcategories, sortOrder]
      properties:
        id: { type: number }
        name: { type: string }
        subcategories:
          type: array
          items: { type: string }
        sortOrder: { type: number }
    CategoryInput:
      type: object
      required: [name]
      properties:
        name: { type: string, minLength: 1 }
        subcategories:
          type: array
          items: { type: string }
        sortOrder: { type: number }
    CategoryUpdate:
      type: object
      properties:
        name: { type: string, minLength: 1 }
        subcategories:
          type: array
          items: { type: string }
        sortOrder: { type: number }
    PlanLine:
      type: object
      required: [id, category, subcategory, planned]
      properties:
        id: { type: number }
        category: { type: string }
        subcategory: { type: string }
        planned: { type: number }
        priority: { type: ["string", "null"] }
        fixedVariable: { type: ["string", "null"] }
        dueDay: { type: ["number", "null"] }
        notes: { type: ["string", "null"] }
    PlanLineInput:
      type: object
      required: [category, subcategory, planned]
      properties:
        category: { type: string, minLength: 1 }
        subcategory: { type: string, minLength: 1 }
        planned: { type: number }
        priority: { type: ["string", "null"] }
        fixedVariable: { type: ["string", "null"] }
        dueDay: { type: ["number", "null"] }
        notes: { type: ["string", "null"] }
    PlanLineUpdate:
      type: object
      properties:
        category: { type: string, minLength: 1 }
        subcategory: { type: string, minLength: 1 }
        planned: { type: number }
        priority: { type: ["string", "null"] }
        fixedVariable: { type: ["string", "null"] }
        dueDay: { type: ["number", "null"] }
        notes: { type: ["string", "null"] }
...
    Rule:
      type: object
      required: [id, pattern, category, subcategory]
      properties:
        id: { type: number }
        pattern: { type: string }
        category: { type: string }
        subcategory: { type: string }
    RuleInput:
      type: object
      required: [pattern, category, subcategory]
      properties:
        pattern: { type: string, minLength: 2 }
        category: { type: string, minLength: 1 }
        subcategory: { type: string, minLength: 1 }
    CategoryActual:
      type: object
      required: [category, planned, actual, remaining]
      properties:
        category: { type: string }
        planned: { type: number }
        actual: { type: number }
        remaining: { type: number }
    DashboardSummary:
      type: object
      required: [month, incomePlanned, incomeActual, plannedExpenses, actualExpenses, remaining, cashFlow, reviewCount, pendingCount, byCategory]
      properties:
        month: { type: string }
        incomePlanned: { type: number }
        incomeActual: { type: number }
        plannedExpenses: { type: number }
        actualExpenses: { type: number }
        remaining: { type: number }
        cashFlow: { type: number }
        reviewCount: { type: number }
        pendingCount: { type: number }
        byCategory:
          type: array
          items:
            $ref: "#/components/schemas/CategoryActual"
    ReconciliationRow:
      type: object
      required: [category, manualTotal, bankTotal, difference, status]
      properties:
        category: { type: string }
        manualTotal: { type: number }
        bankTotal: { type: number }
        difference: { type: number }
        status: { type: string }