# lib/api-zod/src/generated/types/categoryUpdate.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Personal household budget API
 * OpenAPI spec version: 0.1.0
 */

export interface CategoryUpdate {
  /** @minLength 1 */
  name?: string;
  subcategories?: string[];
  sortOrder?: number;
}