# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Personal household budget API
 * OpenAPI spec version: 0.1.0
 */
import type { CategoryActual } from './categoryActual';

export interface DashboardSummary {
  month: string;
  incomePlanned: number;
  incomeActual: number;
  plannedExpenses: number;
  actualExpenses: number;
  remaining: number;
  cashFlow: number;
  reviewCount: number;
  pendingCount: number;
  byCategory: CategoryActual[];
}