# lib/api-zod/src/generated/types/incomeSourceInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Personal household budget API
 * OpenAPI spec version: 0.1.0
 */

export interface IncomeSourceInput {
  /** @minLength 1 */
  name: string;
  /** @nullable */
  owner?: string | null;
  frequency: string;
  netAmount: number;
  monthlyEquivalent: number;
  /** @nullable */
  notes?: string | null;
}