# lib/api-zod/src/generated/types/planLine.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Personal household budget API
 * OpenAPI spec version: 0.1.0
 */

export interface PlanLine {
  id: number;
  category: string;
  subcategory: string;
  planned: number;
  /** @nullable */
  priority?: string | null;
  /** @nullable */
  fixedVariable?: string | null;
  /** @nullable */
  dueDay?: number | null;
  /** @nullable */
  notes?: string | null;
}